Traffic signal recognization/requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
numpy
pandas
matplotlib
scikit-learn
requests
pillow
python-multipart
